visual_inspection_app/requirements.txt
streamlit>=1.49.0
Pillow>=9.0.0
mlflow>=2.0.0
databricks-sdk>=0.12.0
python-dotenv>=1.0.0
openai>=1.0.0
